src/test/java/com/thoughtworks/jpabatch/rest/EmployeeServiceTest.java
package com.thoughtworks.jpabatch.rest;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import static org.mockito.ArgumentMatchers.anyIterable;
import static org.mockito.Mockito.verify;

@RunWith(MockitoJUnitRunner.class)
public class EmployeeServiceTest {

    @Mock
    private EmployeeRepository employeeRepository;

    @InjectMocks
    private EmployeeService employeeService;

    @Test
    public void shouldCreateEmployees() {
        this.employeeService.createEmployees();
        verify(this.employeeRepository).saveAll(anyIterable());
    }
}